requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-telegram-bot==20.6
openai==1.3.5
python-dotenv==1.0.0
pyngrok==7.0.1
httpx==0.25.2
requests==2.31.0
pydantic-settings==2.1.0 